Change signature of `World::main`
