+++ crates/typst-ide/src/lib.rs
@@ -139,8 +139,8 @@ mod tests {
             &self.base.book
         }
 
-        fn main(&self) -> Source {
-            self.main.clone()
+        fn main(&self) -> FileId {
+            self.main.id()
         }
 
         fn source(&self, id: FileId) -> FileResult<Source> {
